Bug fix: address the lane id sign change issue
Occasionally, different OpenDrive roads in the map are stitched to form
a continuous road even if they have opposite road directions;
that is, at such a stitch point, the lane ID could suddenly change its
sign although it is still the same lane in the visual sense.
This function thus uses  just the given initial reference waypoint
rather than each waypoint along the lane to determine  if the current
lane has the same direction as the ego lane.